+++ src/file_util.c
@@ -8,6 +8,7 @@
   * 	E-mail: ste4864@pennwest.edu
   * 
   * Program 1 - fileopen
+  * Group 3
   * CSC 460 - Language Translation
   */
 
@@ -196,8 +197,7 @@ short promptUserOverwriteSelection() {
     return (short) user_pick;
 }
 
-char * promptFilename() {
-
+char * promptInputFilename() {
     char * inputFilename = getString();
 
     if(inputFilename[0] == '\0') {
@@ -208,10 +208,22 @@ char * promptFilename() {
     
 }
 
+char * promptOutputFilename() {
+    char * outputFilename = getString();
+    if(outputFilename[0] == '\0') {
+        printf("No file name was entered. Defaulting to inputfilename.out");
+        outputFilename = (char *) realloc(outputFilename, (sizeof(char) * (strlen(CompFiles.input_file_name) + 4)));
+        strcpy(outputFilename, CompFiles.input_file_name);
+        outputFilename = addExtension(outputFilename, "out");
+
+    }
+    return outputFilename;
+}
+
 char * getString() {
     short max_characters = 49;
     short counter = 0;
-    char * temp_string = (char *) malloc (sizeof(char) * max_characters + 1);
+    char * temp_string = (char *) malloc (sizeof(char) * (max_characters + 1));
     char temp_char;
     
     do{
@@ -228,11 +240,8 @@ char * getString() {
        temp_char != '\0' &&
        counter <= max_characters ); 
 
-    char * new_string = (char *) malloc (sizeof(char) * (counter + 4));
-    strcpy(new_string, temp_string);
-    free(temp_string);
-
-    return new_string;
+    temp_string = (char *) realloc (temp_string, (sizeof(char) * (counter + 4)));
+    return temp_string;
 }
 
 #pragma endregion prompts
@@ -279,12 +288,11 @@ short CompFiles_ValidateInputFile(const char * filename) {
         strcpy(tempfilename, filename);
     } else {
         printf("\nPlease provide an input filename: ");
-        tempfilename = promptFilename();
+        tempfilename = promptInputFilename();
     }
     while(CompFiles.input_file_state != COMPFILES_STATE_NAME_VALIDATED && CompFiles.terminate_requested != 1) {
 
         file_extension_parse = filenameHasExtension(tempfilename);
-
 
         if(file_extension_parse == FILENAME_HAS_NO_PERIOD) {
             printf("\n\t- Your input file has no extension. Defaulting to .in. \n");
@@ -297,21 +305,13 @@ short CompFiles_ValidateInputFile(const char * filename) {
         if(file_extension_parse < 0 ) {
             printf("\n\t- That is not a valid filename. Input a new filename: ");
             free(tempfilename);
-            tempfilename = getString();
-            if(tempfilename[0] == '\0') {
-                printf("\n\t - Terminate program request received.\n");
-                CompFiles.terminate_requested = 1;
-            }
+            tempfilename = promptInputFilename();
         } else {
             short file_exists = fileExists(tempfilename);
             if(!file_exists) {
                 printf("\n\t - That file doesn't exist. Please provide a new filename: ");
                 free(tempfilename);
-                tempfilename = getString();
-                if(tempfilename[0] == '\0') {
-                    printf("\n\t - Terminate program request received.\n");
-                    CompFiles.terminate_requested = 1;
-                }
+                tempfilename = promptInputFilename();
             } else {
                 CompFiles.input_file_state = COMPFILES_STATE_NAME_VALIDATED;
                 CompFiles_LoadInputFile(fopen(tempfilename, "r"));
@@ -319,24 +319,24 @@ short CompFiles_ValidateInputFile(const char * filename) {
             }
         }
     }
+    if(CompFiles.terminate_requested == 1) {
+        printf("\n\t - Terminate program request received.\n");
+    }
     return CompFiles.terminate_requested;
 }
 
 
 
 short CompFiles_ValidateOutputFile(const char * filename){
-    /*short file_extension_parse;*/
-    /* char * tempfilename; 
+    /*
+    short file_extension_parse;
+    char * tempfilename = NULL;
     if(filename != NULL) {
-
         char * tempfilename = malloc(sizeof(char) * (strlen(filename) + 4));
         strcpy(tempfilename, filename);
     } else {
         printf("\nPlease provide an output filename: ");
-        short user_wants_to_exit = promptFilename(tempfilename);
-        if(user_wants_to_exit) {
-            CompFiles.terminate_requested = 1;
-        }
+        promptOutputFilename(tempfilename);
     }
     while(CompFiles.output_file_state != COMPFILES_STATE_NAME_NEEDS_VALIDATION && CompFiles.terminate_requested != 1)
     {
@@ -362,11 +362,13 @@ short CompFiles_ValidateOutputFile(const char * filename){
                 } else {
                     CompFiles.output_file_state = COMPFILES_STATE_NAME_VALIDATED;
                     CompFiles_LoadOutputFile(fopen(tempfilename, "w"));
-                    /*CompFiles.output_file_name = tempfilename;*/
+                    /*CompFiles.output_file_name = tempfilename;
 
-               /* } */
+               }
 
-   /* } */
+    }
+
+ */ 
 }
 
 void CompFiles_LoadInputFile(FILE * newInputFile) {